Move `MontgomeryXpoint` to its own module
